% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\docType{data}
\name{LAD}
\alias{LAD}
\title{'rpart'-method: List of required functions for inducing 'rpart'-like LAD regression trees}
\format{
An object of class \code{list} of length 4.
}
\usage{
LAD
}
\description{
'rpart'-method: List of required functions for inducing 'rpart'-like LAD regression trees
}
\examples{
  mystate <- data.frame(state.x77, region = state.region)
  names(mystate) <- casefold(names(mystate)) #remove mixed case

  fit <-  rpart(murder ~ ., data = mystate, minsplit = 10, method = LAD)
  plot(fit); text(fit)

}
\keyword{datasets}
