% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.LAD-package.R
\docType{package}
\name{rpart.LAD-package}
\alias{rpart.LAD}
\alias{rpart.LAD-package}
\title{rpart.LAD: Least Absolute Deviation Regression Trees}
\description{
Recursive partitioning for least absolute deviation regression trees. Another algorithm from the 1984 book by Breiman, Friedman, Olshen and Stone in addition to the 'rpart' package (Breiman, Friedman, Olshen, Stone (1984, ISBN:9780412048418).
}
\keyword{internal}
